[Perfect Negotiation] Pass Perfect Negotiation WPTs with proper ONN.
This CL migrates to the spec-compliant OnNegotiationNeededEvent[1], only
firing it it the Operations Chain is empty. It also makes the
"negotiationneeded" and "icecandidate" events fire without additional
delay. This is important, because Perfect Negotiation is very
timing-sensitive, and the way we've been firing these events is wrong.

A bug in the Perfect Negotiation WPT is discovered: if an incoming
follow-up offer is sent to the impolite peer while the SRD(answer) of
the previous O/A exchange is still pending, the impolite peer would
incorrectly ignore the offer. This is fixed with the introduction of a
boolean variable to keep track of if SRD is pending, allowing the
impolite peer to predict that it is about to become "stable" and can
thus chain the incoming offer onto its Operations Chain.

With these changes, we pass the Perfect Negotiation WPTs! Hurray!

[1] https://webrtc-review.googlesource.com/c/src/+/180620

Bug: chromium:1060083
Change-Id: I8e8c684ef6cfc1faac0256d33b8fc89be3fc614f